refactor(client): share the way an embedded note follows a link within it
